Stop counting what gets downloaded
A self-hosted downloader has no business keeping a tally of what it served, and
nothing was reading the one we kept. The badge endpoint had no caller: the
frontend's generated client exposed getRequestCount and never invoked it, and
the README's visits badge points at an unrelated service.

Gone with it: GET /api/App/requests, the increment in FetchController,
IRequestCounterService and its implementation, IDataPaths, DATA_PATH, the /data
volume, and the RequestCounter type. ArgonFetch now writes nothing but the
yt-dlp and FFmpeg binaries it fetches at boot.

ArgonFetch.Domain held nothing but that one type, so the project goes too.

The checked-in OpenAPI schema and the generated client were regenerated rather
than hand-edited, which also picks up the /api/Stream/Archive endpoint and the
required url parameter that both had been missing since they were added.
