+++ compose.yml
@@ -11,9 +11,6 @@ services:
       # yt-dlp and FFmpeg are fetched on boot rather than baked into the image. Keeping them
       # here means a restart reuses them instead of downloading roughly 100MB again.
       - tools:/tools
-      # The request counter, and nothing else. There is no database - this is one small JSON
-      # file, kept here so the total survives a restart.
-      - data:/data
       # Uncomment for sources that need a signed-in session, e.g. Instagram, and set
       # COOKIES_PATH=/config/cookies.txt in .env. Treat the file as a credential.
       # - ./cookies.txt:/config/cookies.txt:ro
@@ -21,4 +18,3 @@ services:
 
 volumes:
   tools:
-  data:
