+++ docs/public/openapi.json
@@ -35,19 +35,6 @@
         }
       }
     },
-    "/api/App/requests": {
-      "get": {
-        "tags": [
-          "App"
-        ],
-        "operationId": "GetRequestCount",
-        "responses": {
-          "200": {
-            "description": "OK"
-          }
-        }
-      }
-    },
     "/api/Fetch/GetResource": {
       "get": {
         "tags": [
@@ -58,6 +45,7 @@
           {
             "name": "url",
             "in": "query",
+            "required": true,
             "schema": {
               "type": "string"
             }
@@ -226,6 +214,99 @@
                 }
               }
             }
+          }
+        }
+      }
+    },
+    "/api/Stream/Archive": {
+      "get": {
+        "tags": [
+          "Stream"
+        ],
+        "operationId": "Archive",
+        "parameters": [
+          {
+            "name": "url",
+            "in": "query",
+            "required": true,
+            "schema": {
+              "type": "string"
+            }
+          },
+          {
+            "name": "jobId",
+            "in": "query",
+            "schema": {
+              "type": "string"
+            }
+          }
+        ],
+        "responses": {
+          "200": {
+            "description": "OK"
+          },
+          "400": {
+            "description": "Bad Request",
+            "content": {
+              "text/plain": {
+                "schema": {
+                  "$ref": "#/components/schemas/ProblemDetails"
+                }
+              },
+              "application/json": {
+                "schema": {
+                  "$ref": "#/components/schemas/ProblemDetails"
+                }
+              },
+              "text/json": {
+                "schema": {
+                  "$ref": "#/components/schemas/ProblemDetails"
+                }
+              }
+            }
+          },
+          "404": {
+            "description": "Not Found",
+            "content": {
+              "text/plain": {
+                "schema": {
+                  "$ref": "#/components/schemas/ProblemDetails"
+                }
+              },
+              "application/json": {
+                "schema": {
+                  "$ref": "#/components/schemas/ProblemDetails"
+                }
+              },
+              "text/json": {
+                "schema": {
+                  "$ref": "#/components/schemas/ProblemDetails"
+                }
+              }
+            }
+          }
+        }
+      }
+    },
+    "/api/Stream/Archive/Progress/{jobId}": {
+      "get": {
+        "tags": [
+          "Stream"
+        ],
+        "operationId": "ArchiveProgress",
+        "parameters": [
+          {
+            "name": "jobId",
+            "in": "path",
+            "required": true,
+            "schema": {
+              "type": "string"
+            }
+          }
+        ],
+        "responses": {
+          "200": {
+            "description": "OK"
           }
         }
       }
@@ -463,6 +544,10 @@
         "properties": {
           "type": {
             "$ref": "#/components/schemas/MediaType"
+          },
+          "requestedUrl": {
+            "type": "string",
+            "nullable": true
           },
           "title": {
             "type": "string",
