+++ src/ArgonFetch.API/Controllers/FetchController.cs
@@ -12,14 +12,12 @@ namespace ArgonFetch.API.Controllers
     public class FetchController : ControllerBase
     {
         private readonly IMediator _mediator;
-        private readonly IRequestCounterService _requestCounter;
         private readonly ILogger<FetchController> _logger;
         private readonly IMaintenanceState _maintenance;
 
-        public FetchController(IMediator mediator, IRequestCounterService requestCounter, ILogger<FetchController> logger, IMaintenanceState maintenance)
+        public FetchController(IMediator mediator, ILogger<FetchController> logger, IMaintenanceState maintenance)
         {
             _mediator = mediator;
-            _requestCounter = requestCounter;
             _logger = logger;
             _maintenance = maintenance;
         }
@@ -51,10 +49,6 @@ namespace ArgonFetch.API.Controllers
             try
             {
                 var result = await _mediator.Send(new GetMediaQuery(url));
-
-                // Counted only on success, so the number reflects media actually served
-                // rather than every malformed URL someone pasted.
-                await _requestCounter.IncrementAsync();
 
                 return Ok(result);
             }
