+++ src/ArgonFetch.Infrastructure/ArgonFetch.Infrastructure.csproj
@@ -11,7 +11,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\ArgonFetch.Domain\ArgonFetch.Domain.csproj" />
     <ProjectReference Include="..\ArgonFetch.Application\ArgonFetch.Application.csproj" />
   </ItemGroup>
