+++ src/ArgonFetch.Tests/ArgonFetch.Tests.csproj
@@ -26,7 +26,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\ArgonFetch.Application\ArgonFetch.Application.csproj" />
-    <ProjectReference Include="..\ArgonFetch.Infrastructure\ArgonFetch.Infrastructure.csproj" />
   </ItemGroup>
 
 </Project>
