+++ template.development.env
@@ -1,10 +1,6 @@
 # Development Configuration Example
 # Copy this to .env for local development
 
-# Where the request counter is kept. There is no database; this is one small JSON file,
-# and leaving it unset writes it next to the built app.
-# DATA_PATH=./data
-
 # ASP.NET Core Configuration
 ASPNETCORE_ENVIRONMENT=Development
 ASPNETCORE_URLS=http://localhost:5000
